Breaking command helper into seperate classes
